chore!: move the npm packages from global to backend; test: create first router test
