+++ frontend/src/httpClient/axios.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 // const base_url = process.env.REACT_APP_BACKEND_API_HOST;
-const base_url = "http://localhost:3001";
+const base_url = "http://localhost:3005";
 
 export const getMatches = () => {
   return axios.get(`${base_url}/match`);
