+++ package.json
@@ -18,7 +18,6 @@
   },
   "homepage": "https://github.com/jonbrook/Euro-2021-Sweepstake#readme",
   "devDependencies": {
-    "@testing-library/user-event": "^13.1.9",
     "eslint": "^7.28.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^8.3.0",
@@ -26,17 +25,5 @@
     "eslint-plugin-prettier": "^3.4.0",
     "prettier": "^2.3.1"
   },
-  "dependencies": {
-    "axios": "^0.21.1",
-    "cors": "^2.8.5",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "lodash": "^4.17.21",
-    "morgan": "^1.10.0",
-    "nanoid": "^3.1.23",
-    "pg": "^8.6.0",
-    "pg-hstore": "^2.3.3",
-    "redis": "^2.8.0",
-    "sequelize": "^6.6.2"
-  }
+  "dependencies": {}
 }
